feat(j-s): rename HearingArrangements to DefentantParty
