+++ apps/judicial-system/web/pages/domur/akaera/malflytjendur/[id].ts
@@ -1,0 +1,3 @@
+import DefentantParty from '@island.is/judicial-system-web/src/routes/Court/Indictments/DefentantParty/DefentantParty'
+
+export default DefentantParty
