--- apps/judicial-system/web/pages/domur/akaera/thingfesting/[id].ts
@@ -1,3 +1,0 @@
-import HearingArrangements from '@island.is/judicial-system-web/src/routes/Court/Indictments/HearingArrangements/HearingArrangements'
-
-export default HearingArrangements
