+++ libs/judicial-system/consts/src/lib/consts.ts
@@ -227,9 +227,8 @@ export const INDICTMENTS_COURT_OVERVIEW_ROUTE = '/domur/akaera/yfirlit'
 export const INDICTMENTS_RECEPTION_AND_ASSIGNMENT_ROUTE =
   '/domur/akaera/mottaka'
 export const INDICTMENTS_SUBPOENA_ROUTE = '/domur/akaera/fyrirkall'
+export const INDICTMENTS_DEFENTANT_PARTY_ROUTE = '/domur/akaera/malflytjendur'
 export const INDICTMENTS_COURT_RECORD_ROUTE = '/domur/akaera/thingbok'
-export const INDICTMENTS_HEARING_ARRANGEMENTS_ROUTE =
-  '/domur/akaera/thingfesting'
 /* COURT ROUTES END */
 
 // Feedback
